+++ types/bsmap/save.ts
@@ -1,11 +1,17 @@
 import { IOptimizeOptionsDifficulty, IOptimizeOptionsInfo } from './optimize.ts';
 import { IBaseOptions } from './options.ts';
 
+export interface ISaveValidate {
+    enabled: boolean;
+    reparse?: boolean;
+}
+
 export interface ISaveOptionsInfo extends IBaseOptions {
     /** Set info destination file path/name. */
     filePath?: string;
     format?: number;
     optimize?: IOptimizeOptionsInfo;
+    validate?: ISaveValidate;
 }
 
 export interface ISaveOptionsDifficulty extends IBaseOptions {
@@ -16,8 +22,9 @@ export interface ISaveOptionsDifficulty extends IBaseOptions {
     filePath?: string;
     format?: number;
     optimize?: IOptimizeOptionsDifficulty;
+    validate?: ISaveValidate;
     dataCheck?: {
-        enable: boolean;
+        enabled: boolean;
         throwError?: boolean;
     };
 }
@@ -25,8 +32,9 @@ export interface ISaveOptionsDifficulty extends IBaseOptions {
 export interface ISaveOptionsDifficultyList extends IBaseOptions {
     format?: number;
     optimize?: IOptimizeOptionsDifficulty;
+    validate?: ISaveValidate;
     dataCheck?: {
-        enable: boolean;
+        enabled: boolean;
         throwError?: boolean;
     };
 }